+++ config.php
@@ -5,6 +5,8 @@ use humhub\modules\verified\Events;
 use humhub\modules\content\widgets\ContainerProfileHeader;
 use humhub\modules\user\widgets\PeopleCard;
 use humhub\modules\space\widgets\SpaceDirectoryCard;
+use humhub\modules\content\widgets\stream\WallStreamEntryWidget;
+use humhub\modules\comment\widgets\Comment;
 
 return [
     'id' => 'verified',
@@ -26,5 +28,15 @@ return [
 			'event' => SpaceDirectoryCard::EVENT_CREATE,
 			'callback' => [Events::class, 'onSpaceDirectoryCardBeforeRun']
 		],
+        [
+			'class' => WallStreamEntryWidget::class,
+			'event' => WallStreamEntryWidget::EVENT_CREATE,
+			'callback' => [Events::class, 'onWallStreamEntryWidgetBeforeRun']
+        ],
+        [
+			'class' => Comment::class,
+			'event' => Comment::EVENT_CREATE,
+			'callback' => [Events::class, 'onCommentBeforeRun']
+		]
     ],
 ];