+++ messages/de/base.php
@@ -1,0 +1,6 @@
+<?php
+
+return array(
+	'Verified User' => 'Verifizierter Benutzer',
+	'Verified Space' => 'Verifizierter Space',
+);